+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 
 ### Changed
 - prevent zoom, if marker is clicked
+- district and region filters are now global filters (not per layer)